Move get_ffi_func to utils
