+++ py-polars/polars/internals/series/utils.py
@@ -4,9 +4,11 @@ import sys
 from typing import TYPE_CHECKING, Any, Callable
 
 import polars.internals as pli
+from polars.datatypes import DataType, dtype_to_ffiname
 
 if TYPE_CHECKING:
     from polars.internals.type_aliases import Namespace
+    from polars.polars import PySeries
 
     if sys.version_info >= (3, 10):
         from typing import ParamSpec
@@ -43,3 +45,32 @@ def expr(
         return wrapper  # type: ignore[return-value]
 
     return decorator
+
+
+def get_ffi_func(
+    name: str,
+    dtype: type[DataType],
+    obj: PySeries,
+) -> Callable[..., Any] | None:
+    """
+    Dynamically obtain the proper ffi function/ method.
+
+    Parameters
+    ----------
+    name
+        function or method name where dtype is replaced by <>
+        for example
+            "call_foo_<>"
+    dtype
+        polars dtype.
+    obj
+        Object to find the method for.
+
+    Returns
+    -------
+    ffi function, or None if not found
+
+    """
+    ffi_name = dtype_to_ffiname(dtype)
+    fname = name.replace("<>", ffi_name)
+    return getattr(obj, fname, None)
